с настройками postgres

+++ src/main/java/sweater/repos/MessageRepo.java
@@ -3,5 +3,5 @@ package sweater.repos;
 import org.springframework.data.repository.CrudRepository;
 import sweater.domain.Message;
 
-public interface MessageRepo extends CrudRepository <Message, Integer> {
+public interface MessageRepo extends CrudRepository <Message, Long> {
 }
